# docs/swagger.yaml
basePath: /
definitions:
...
  models.ApplicationSettings:
    properties:
      analytics_code:
        description: Google Analytics kodu
        type: string
      contact_info:
        additionalProperties:
          type: string
        description: İletişim bilgileri (telefon, e-posta vb.)
        type: object
      default_lang:
        description: Varsayılan dil
        type: string
      description:
        additionalProperties:
          type: string
        description: Çok dilli açıklama
        type: object
      favicon_url:
        description: Favicon URL'si
        type: string
      id:
        type: string
      logo_url:
        description: Logo URL'si
        type: string
      maintenance_mode:
        type: boolean
      maintenance_msg:
        additionalProperties:
          type: string
        description: Çok dilli bakım modu mesajı
        type: object
      meta_tags:
        additionalProperties:
          $ref: '#/definitions/models.MetaTag'
        description: SEO meta bilgileri
        type: object
      social_media:
        additionalProperties:
          $ref: '#/definitions/models.SocialMedia'
        description: SocialMedia     map[string]string  `bson:"social_media" json:"social_media"`
          // Sosyal medya bağlantıları
        type: object
      supported_langs:
        description: Desteklenen diller
        items:
          type: string
        type: array
      title:
        additionalProperties:
          type: string
        description: Çok dilli başlık
        type: object
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.MetaTag:
    properties:
      description:
        type: string
      keywords:
        description: Dizi olarak tanımlandı
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      permissions:
        description: Permission referansları
        items:
          type: string
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    required:
    - name
    type: object
...
  models.SocialMedia:
    properties:
      active:
        description: Aktif mi?
        type: boolean
      name:
        description: 'Örnek: Facebook, Twitter'
        example: Facebook
        type: string
      'url example:':
        description: 'Örnek: https://facebook.com/yourpage'
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the admin panel project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Panel API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Retrieve all roles with their permissions and details
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
        "500":
          description: Failed to retrieve roles
          schema:
            additionalProperties: true
            type: object
      summary: Get all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Add a new role with its permissions and details
      parameters:
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create role
          schema:
            additionalProperties: true
            type: object
      summary: Create a new role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Remove a role by its unique identifier
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Role deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete role
          schema:
            additionalProperties: true
            type: object
      summary: Delete a role
      tags:
      - Roles
    get:
      description: Retrieve a role with its permissions and details by its unique
        identifier
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details
          schema:
            $ref: '#/definitions/models.Role'
        "400":
          description: Invalid role ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve role
          schema:
            additionalProperties: true
            type: object
      summary: Get a role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role details, including permissions
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role ID or request payload
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update role
          schema:
            additionalProperties: true
            type: object
      summary: Update a role
      tags:
      - Roles
  /settings:
    get:
      description: Retrieve all application settings including general settings and
        configurations
      produces:
      - application/json
      responses:
        "200":
          description: Application settings
          schema:
            $ref: '#/definitions/models.ApplicationSettings'
        "500":
          description: Failed to retrieve settings
          schema:
            additionalProperties: true
            type: object
      summary: Get application settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update application settings with new values
      parameters:
      - description: Updated settings data
        in: body
        name: update
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Settings updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update settings
          schema:
            additionalProperties: true
            type: object
      summary: Update application settings
      tags:
      - Settings
  /settings/social-media:
    get:
      description: Retrieve all social media links configured for the application
      produces:
      - application/json
      responses:
        "200":
          description: List of social media links
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve social media links
          schema:
            additionalProperties: true
            type: object
      summary: Get social media links
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Update the social media links for the application
      parameters:
      - description: Updated social media links
        in: body
        name: links
        required: true
        schema:
          additionalProperties:
            $ref: '#/definitions/models.SocialMedia'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Social media links updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update social media links
          schema:
            additionalProperties: true
            type: object
      summary: Update social media links
      tags:
      - Social Media
...
security:
- BearerAuth: []
swagger: "2.0"